[ACA-1060] [Recent Files] Working copy is incorrectly displayed for a file locked for offline editing
-moved common code to PageComponent to avoid duplication
-used the imageResolver and onNodeSelect on recent files also
