+++ src/app/components/files/files.component.html
@@ -101,7 +101,7 @@
                 [imageResolver]="imageResolver"
 
                 (node-dblclick)="onNodeDoubleClick($event)"
-                (node-select)="onNodeSelect($event)">
+                (node-select)="onNodeSelect($event, documentList)">
 
                 <data-columns>
                     <data-column
