+++ src/app/components/files/files.component.ts
@@ -16,14 +16,13 @@
  */
 
 import { Observable, Subscription } from 'rxjs/Rx';
-import { Component, OnInit, OnDestroy, ViewChild, NgZone } from '@angular/core';
+import { Component, OnInit, OnDestroy, NgZone } from '@angular/core';
 import { Router, ActivatedRoute, Params } from '@angular/router';
 import { MinimalNodeEntity, MinimalNodeEntryEntity, PathElementEntity, NodePaging, PathElement } from 'alfresco-js-api';
 import {
     UploadService, FileUploadEvent, NodesApiService,
     ContentService, AlfrescoApiService, UserPreferencesService
 } from '@alfresco/adf-core';
-import { DocumentListComponent, ShareDataRow } from '@alfresco/adf-content-services';
 
 import { BrowsingFilesService } from '../../common/services/browsing-files.service';
 import { ContentManagementService } from '../../common/services/content-management.service';
@@ -35,8 +34,6 @@ import { PageComponent } from '../page.component';
     templateUrl: './files.component.html'
 })
 export class FilesComponent extends PageComponent implements OnInit, OnDestroy {
-    @ViewChild(DocumentListComponent)
-    documentList: DocumentListComponent;
 
     private routeData: any = {};
     isValidPath = true;
@@ -139,7 +136,7 @@ export class FilesComponent extends PageComponent implements OnInit, OnDestroy {
                     this.navigate(node.id);
                 }
 
-                if (node.isLocked) {
+                if (PageComponent.isLockedNode(node)) {
                     event.preventDefault();
 
                 } else if (node.isFile) {
@@ -147,33 +144,6 @@ export class FilesComponent extends PageComponent implements OnInit, OnDestroy {
                 }
             }
 
-        }
-    }
-
-    onNodeSelect(event) {
-        if (!!event.detail && !!event.detail.node) {
-
-            const node: MinimalNodeEntryEntity = event.detail.node.entry;
-            if (node && node.isLocked) {
-                this.unSelectLockedNodes();
-            }
-        }
-    }
-
-    unSelectLockedNodes() {
-        this.documentList.selection = this.documentList.selection.filter(item => !item.entry.isLocked);
-
-        const dataTable = this.documentList.dataTable;
-        if (dataTable && dataTable.data) {
-            const rows = dataTable.data.getRows();
-
-            if (rows && rows.length > 0) {
-                rows.forEach(r => {
-                    if (r.getValue('isLocked')) {
-                        r.isSelected = false;
-                    }
-                });
-            }
         }
     }
 
@@ -307,14 +277,5 @@ export class FilesComponent extends PageComponent implements OnInit, OnDestroy {
             return this.node.path.elements[0].id === nodeId;
         }
         return false;
-    }
-
-    imageResolver(row: ShareDataRow): string | null {
-        const entry: MinimalNodeEntryEntity = row.node.entry;
-
-        if (entry.isLocked) {
-            return '/assets/images/ic_lock_black_24dp_1x.png';
-        }
-        return null;
     }
 }
